pyproject.toml
[project]
name = "irontrack"
version = "0.1.0"
description = "FastAPI web application for tracking workouts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "advanced-alchemy[uuid]>=1.4.5",
    "anyio>=4.9.0",
    "asgi-correlation-id>=4.3.4",
    "asyncpg>=0.30.0",
    "click>=8.2.1",
    "email-validator>=2.2.0",
    "fastapi>=0.116.1",
    "fastapi-cache2>=0.2.2",
    "fastapi-cli>=0.0.8",
    "msgspec>=0.19.0",
    "pwdlib[argon2]>=0.2.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]
lint = [
    "mypy>=1.18.2",
    "ruff>=0.13.2",
]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".idea",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "__init__.py",
    "src/db/alembic/*",
]
fix = true
line-length = 120
indent-width = 4
lint.fixable = ["ALL"]
lint.ignore = [
    "SIM117",  # multiple-with-statements
    "COM812",  # missing-trailing-comma
    "D100",    # pydocstyle - missing docstring in public module
    "D101",    # pydocstyle - missing docstring in public class
    "D102",    # pydocstyle - missing docstring in public method
    "D103",    # pydocstyle - missing docstring in public function
    "D104",    # pydocstyle - missing docstring in public package
    "D105",    # pydocstyle - missing docstring in magic method
    "D106",    # pydocstyle - missing docstring in public nested class
    "D107",    # pydocstyle - missing docstring in __init__
    "S106",    # hardcoded-password-func-arg
    "ANN401",  # any-type
    "ARG001",  # unused-function-argument
]
lint.select = ["ALL"]
lint.extend-select = ["I"]
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
src = ["src", "tests"]
target-version = "py312"

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.mccabe]
max-complexity = 8

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
show_error_codes = true
disallow_any_unimported = false
disallow_subclassing_any = true
warn_unreachable = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = '''(?x)(
    ^.venv/
  | ^src/db/alembic/
  | ^src/scripts/
)'''

[[tool.mypy.overrides]]
warn_return_any = false
module = [
    "src.lib.coders.*",
    "src.lib.json_response.*",
    "src.lib.jwt_utils.*"
]

[[tool.mypy.overrides]]
disallow_subclassing_any = false
module = [
    "src.lib.coders.*"
]